% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{den.est}
\alias{den.est}
\title{Kernel density estimation results at boundary}
\usage{
den.est(u, ui, h)
}
\arguments{
\item{u}{the point to be estimated}

\item{ui}{input samples of the target density function}

\item{h}{the bandwidth (smoothing parameter)}
}
\value{
a vector of estimated density function values at the points to be estimated
}
\description{
Kernel density estimation results using linear multiple kernel of Generlized Jacknifing method
}
\examples{
\dontrun{
x <- seq(0, 8, by = 0.02)
y <- den.est(x, dat, 2 * bw.bcv(dat))
y0 <- den.est(0, dat, 2 * bw.bcv(dat))
cat("The estimated value at boundary 0 by using GJack is ", y0)
}
}
