% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{kernel.new}
\alias{kernel.new}
\title{linear multiple kernel function for Generlized Jacknifing method of KDE}
\usage{
kernel.new(x, u, h)
}
\arguments{
\item{x}{the input}

\item{u}{the input}

\item{h}{the bandwidth (smoothing parameter)}
}
\value{
function value(s) of the kernel function at certain point(s)
}
\description{
This new kernel function is special kernel function form of Generlized Jacknifing method
}
\examples{
\dontrun{
set.seed(123)
library(stats)
x <- seq(0, 8, by = 0.02)
dat <- rexp(1000, 1)
h <- 2 * bw.bcv(dat)
kernel.new.values <- kernel.new((x-dat)/h, x, h)
}
}
