% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{count}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance of the normal distribution}

\item{x0}{the initial value}

\item{N}{number of trials}
}
\value{
a vector
}
\description{
count the number of rejected
}
