\name{StatComp20024-package}
\alias{StatComp20024-package}
\alias{StatComp20024}
\docType{package}
\title{
  Statistics Computation Course Final Project
}
\description{
  Statistics Computation Course Final Project
  Include: 4 individual functions with detail information and example and all homeworks with rmd and html.
}
\details{
  This section should provide a more detailed overview of how to use the package, including the most important functions.
  
  1. An KDE method descripted by "Liu, H., Lafferty, J. and Wasserman, L., 2007, March. Sparse nonparametric density estimation in high dimensions using the rodeo. In Artificial Intelligence and Statistics, 283-290."
  
  2. A boundary correction approach in KDE method descripted by "Jones M. C., 1993. Nonlinear Solution of Linear Inverse Problems by Wavelet-Vaguelette Decomposition. Statistics and
Computing 3(1993), 135-146."
}
\author{
Jiayang Zhao, Zhaojy19@mail.ustc.edu.cn.

Maintainer: Jiayang Zhao <Zhaojy19@mail.ustc.edu.cn>
}
\references{
  1. Liu, H., Lafferty, J. and Wasserman, L., 2007, March. Sparse nonparametric density estimation in high dimensions using the rodeo. In Artificial Intelligence and Statistics (pp. 283-290)
  
  2. Jones M. C., 1993. Nonlinear Solution of Linear Inverse Problems by Wavelet-Vaguelette Decomposition. Statistics and
Computing 3(1993), 135-146.
}
\keyword{package, computation, rodeo, KDE, boundary correction}
\seealso{
  https://github.com/zhaojiayang96
}
